tweaks to mosquito_leech_relocate
this is particularly complex because I do not feel it should be
necessary to catch incomplete instances of Relocation where leech_from
is omitted.  When omitted, the class __init__ should deduce instead that
it is not a leech, not assign it to be leech_relo but just relo--and as
a result, fail in the relocation logic branch, instead of the leech_relo
branch.